tests/submission/Ms2ger/dynamic-markup-insertion/innerhtml-03.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>innerHTML in XHTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="http://www.whatwg.org/html5/#dom-innerhtml"/>
<link rel="help" href="http://www.whatwg.org/html5/#xml-fragment-serialization-algorithm"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<p>Note: this test relies on the user agent using the '"' character to quote attributes, which is not currently required by the specification.</p>
<div id="log"></div>
<script><![CDATA[
test(function() {
  var el = document.createElement("div");
  el.appendChild(document.createElement("xmp"))
    .appendChild(document.createElement("span"))
    .appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "<xmp xmlns=\"http://www.w3.org/1999/xhtml\"><span>&lt;<\/span><\/xmp>");
})
test(function() {
  var el = document.createElement("xmp");
  el.appendChild(document.createElement("span"))
    .appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "<span xmlns=\"http://www.w3.org/1999/xhtml\">&lt;<\/span>");
})
test(function() {
  var el = document.createElement("xmp");
  el.appendChild(document.createTextNode("<"));
  assert_equals(el.innerHTML, "&lt;");
})
test(function() {
  var el = document.createElement("div");
  el.appendChild(document.createElement("br"));
  assert_equals(el.innerHTML, "<br xmlns=\"http://www.w3.org/1999/xhtml\" />");
})
]]></script>
</body>
</html>
